snap: Filter out the SNAP variables by their contents
Use the upstream-proposed patch to cleanup the snap variables without
leaking any snap-related information to the terminal children.

As per this we can avoid (and actually we must not) do the filtering in
the launcher since it must provide everything to the binary so that
it's able to clean things up.